+++ app/Http/Controllers/Admin/UserManagement/ModuleController.php
@@ -3,6 +3,8 @@
 namespace App\Http\Controllers\Admin\UserManagement;
 
 use App\Http\Controllers\Controller;
+use App\Http\Requests\Admin\UserManagement\Module\ModuleStoreRequest;
+use App\Http\Requests\Admin\UserManagement\Module\ModuleUpdateRequest;
 use App\Models\Module;
 use Illuminate\Http\Request;
 
@@ -12,6 +14,9 @@ class ModuleController extends Controller
 
     public function index()
     {
+        if (checkPermission('display-module') == true) {
+            return redirect()->route('noAccess');
+        }
         return view('admin.user-management.modules.index', [
             'modules' => Module::latest()->paginate(10),
         ]);
@@ -20,22 +25,31 @@ class ModuleController extends Controller
 
     public function create()
     {
+        if (checkPermission('create-module') == true) {
+            return redirect()->route('noAccess');
+        }
         return view('admin.user-management.modules.create');
     }
 
 
-    public function store(Request $request){
+    public function store(ModuleStoreRequest $request){
 
+        if (checkPermission('create-module') == true) {
+            return redirect()->route('noAccess');
+        }
         Module::create([
             'title' => $request->title,
         ]);
 
-        return redirect()->route('admin.modules.index')->with('success', 'Module create done');
+        return redirect()->route('admin.modules.index')->with('success', 'Recode Create Successfully!');
     }
 
 
     public function edit($id)
     {
+        if (checkPermission('edit-module') == true) {
+            return redirect()->route('noAccess');
+        }
         $data = Module::find($id);
 
         return view('admin.user-management.modules.edit', [
@@ -44,23 +58,29 @@ class ModuleController extends Controller
     }
 
 
-    public function update(Request $request, $id){
+    public function update(ModuleUpdateRequest $request, $id){
 
+        if (checkPermission('edit-module') == true) {
+            return redirect()->route('noAccess');
+        }
         Module::find($id)->update([
             'title' => $request->title,
         ]);
 
-        return redirect()->route('admin.modules.index')->with('success', 'Module Update done');
+        return redirect()->route('admin.modules.index')->with('success', 'Recode Update Successfully!');
     }
 
 
     public function destroy($id){
 
+        if (checkPermission('delete-module') == true) {
+            return redirect()->route('noAccess');
+        }
         $data = Module::find($id);
 
         $data->delete();
 
-        return redirect()->route('admin.modules.index')->with('success', 'Module delete done');
+        return redirect()->route('admin.modules.index')->with('success', 'Recode Delete Successfully!');
     }
