+++ app/Http/Controllers/Admin/UserManagement/UserController.php
@@ -3,6 +3,8 @@
 namespace App\Http\Controllers\Admin\UserManagement;
 
 use App\Http\Controllers\Controller;
+use App\Http\Requests\Admin\UserManagement\User\UserStoreRequest;
+use App\Http\Requests\Admin\UserManagement\User\UserUpdateRequest;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
@@ -11,28 +13,34 @@ use Spatie\Permission\Models\Permission;
 
 class UserController extends Controller
 {
-
     public function index()
     {
-        if(checkPermission('edit articles') == true){
-            return redirect()->route("noAccess");
+
+        if (checkPermission('display-user') == true) {
+            return redirect()->route('noAccess');
         }
 
         return view('admin.user-management.users.index', [
             'users' => User::latest()->paginate(10),
-            'roles' => Role::all(),
+            // 'roles' => Role::all(),
         ]);
     }
 
-
     public function create()
     {
+        if (checkPermission('create-user') == true) {
+            return redirect()->route('noAccess');
+        }
         return view('admin.user-management.users.create', [
             'roles' => Role::all(),
         ]);
     }
 
-    public function store(Request $request){
+    public function store(UserStoreRequest $request)
+    {
+         if (checkPermission('create-user') == true) {
+            return redirect()->route('noAccess');
+        }
 
         $new_user = User::create([
             'name' => $request->name,
@@ -42,14 +50,20 @@ class UserController extends Controller
         ]);
 
         $role = Role::findById($request->role_id);
+
         $user = User::find($new_user->id);
         $user->assignRole($role->name);
 
-        return redirect()->route('admin.users.index')->with('success', 'User create done');
+        return redirect()
+            ->route('admin.users.index')
+            ->with('success', 'User create done');
     }
 
     public function edit($id)
     {
+         if (checkPermission('edit-user') == true) {
+            return redirect()->route('noAccess');
+        }
         $data = User::find($id);
         return view('admin.user-management.users.edit', [
             'roles' => Role::all(),
@@ -57,12 +71,17 @@ class UserController extends Controller
         ]);
     }
 
-    public function update(Request $request, $id){
-        $user = User::find($id);
+    public function update(UserUpdateRequest $request, $id)
+    {
+         if (checkPermission('edit-user') == true) {
+            return redirect()->route('noAccess');
+        }
+
+        $user = User::findOrFail($id);
 
         $old_role_id = $user->role_id;
 
-        if($old_role_id != $request->role_id){
+        if ($old_role_id != $request->role_id) {
             $user->removeRole($old_role_id);
 
             $role = Role::findById($request->role_id);
@@ -76,20 +95,22 @@ class UserController extends Controller
             'role_id' => $request->role_id,
         ]);
 
-
-
-        return redirect()->route('admin.users.index')->with('success', 'User Update done');
+        return redirect()
+            ->route('admin.users.index')
+            ->with('success', 'User Update done');
     }
 
-
-    public function destroy($id){
-
-        $data = User::find($id);
+    public function destroy($id)
+    {
+         if (checkPermission('delete-user') == true) {
+            return redirect()->route('noAccess');
+        }
+        $data = User::findOrFail($id);
 
         $data->delete();
 
-        return redirect()->route('admin.users.index')->with('success', 'User delete done');
+        return redirect()
+            ->route('admin.users.index')
+            ->with('success', 'User delete done');
     }
-
-
 }
